Extract MVR scene node reading
