+++ mvr/mvrimporter.h
@@ -25,37 +25,7 @@
 
 class wxInputStream;
 
-#include "mvrscene.h"
-
-enum class MvrImportMode {
-    ReplaceProject,
-    ParseOnly,
-};
-
-enum class MvrImportSourceKind {
-    ExternalImport,
-    ProjectRestore,
-    MergeImport,
-};
-
-struct MvrImportOptions {
-    bool promptConflicts = true;
-    bool applyDictionary = true;
-    bool preserveMvrGdtfReferences = true;
-    bool allowDummyFallback = true;
-    MvrImportSourceKind sourceKind = MvrImportSourceKind::ExternalImport;
-};
-
-struct MvrImportDiagnostic {
-    std::string code;
-    std::string message;
-};
-
-struct MvrImportResult {
-    MvrScene scene;
-    std::unordered_map<std::string, std::string> fixtureUuidRemap;
-    std::vector<MvrImportDiagnostic> diagnostics;
-};
+#include "mvr_import_types.h"
 
 // Responsible for importing .mvr files into the application's internal data model
 class MvrImporter
